Number pairs components (w/o tests):     -- interface     -- generic component     -- specified extensions for Integer, Long and Float numbers     -- data models for all listed above     -- minor package structure refactoring
