+++ components-impl/src/main/java/ru/rougegibbons/components/interfaces/core/StringComponent.java
@@ -1,6 +1,7 @@
-package ru.rougegibbons.components.interfaces;
+package ru.rougegibbons.components.interfaces.core;
 
 import org.jetbrains.annotations.NotNull;
+import ru.rougegibbons.components.interfaces.Component;
 
 /**
  * {@link Component} interface extension for text containers.
